backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
supabase==2.10.0
stripe==11.4.0
python-jose[cryptography]==3.3.0
asyncpg==0.30.0
slowapi==0.1.9
python-multipart==0.0.20
structlog==25.1.0
